# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "0to1log API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/cron/news-pipeline": {
      "post": {
        "tags": [
          "cron"
        ],
        "summary": "Run News Pipeline",
        "description": "Trigger daily news pipeline. Returns 202 immediately; runs async.",
        "operationId": "run_news_pipeline_api_cron_news_pipeline_post",
        "parameters": [
          {
            "name": "x-cron-secret",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Cron-Secret"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineAcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid cron secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/drafts": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Drafts",
        "description": "List all posts with status='draft'.",
        "operationId": "list_drafts_api_admin_drafts_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDraftListItem"
                  },
                  "title": "Response List Drafts Api Admin Drafts Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/drafts/{slug}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Draft",
        "description": "Get a single draft by slug with full detail.",
        "operationId": "get_draft_api_admin_drafts__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDraftDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Draft not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/posts/{post_id}/publish": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Publish Post",
        "description": "Change post status from draft to published.",
        "operationId": "publish_post_api_admin_posts__post_id__publish_patch",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPublishResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/posts/{post_id}/update": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Post",
        "description": "Update draft content fields.",
        "operationId": "update_post_api_admin_posts__post_id__update_patch",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDraftDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "timestamp"
        ],
        "title": "HealthResponse",
        "description": "GET /health response."
      },
      "PipelineAcceptedResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "message"
        ],
        "title": "PipelineAcceptedResponse",
        "description": "POST /cron/news-pipeline 202 response."
      },
      "PostDraftDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "post_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en"
          },
          "content_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Original"
          },
          "has_news": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has News"
          },
          "no_news_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "No News Notice"
          },
          "recent_fallback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Fallback"
          },
          "content_beginner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Beginner"
          },
          "content_learner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Learner"
          },
          "content_expert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Expert"
          },
          "guide_items": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptGuideItems"
              },
              {
                "type": "null"
              }
            ]
          },
          "related_news": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedNews"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Urls",
            "default": []
          },
          "news_temperature": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Temperature"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "pipeline_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Batch Id"
          },
          "translation_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Group Id"
          },
          "source_post_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Post Id"
          },
          "source_post_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Post Version"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "category",
          "status"
        ],
        "title": "PostDraftDetail",
        "description": "Draft detail endpoint response \u2014 full view."
      },
      "PostDraftListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "post_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "news_temperature": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Temperature"
          },
          "pipeline_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Batch Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "category",
          "status"
        ],
        "title": "PostDraftListItem",
        "description": "Draft list endpoint response item \u2014 summary view."
      },
      "PostPublishResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "status",
          "published_at"
        ],
        "title": "PostPublishResponse",
        "description": "PATCH /admin/posts/{id}/publish response."
      },
      "PostUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Original"
          },
          "content_beginner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Beginner"
          },
          "content_learner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Learner"
          },
          "content_expert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Expert"
          },
          "guide_items": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptGuideItems"
              },
              {
                "type": "null"
              }
            ]
          },
          "related_news": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedNews"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "PostUpdateRequest",
        "description": "PATCH /admin/posts/{id}/update request body."
      },
      "PromptGuideItems": {
        "properties": {
          "one_liner": {
            "type": "string",
            "title": "One Liner"
          },
          "action_item": {
            "type": "string",
            "title": "Action Item"
          },
          "critical_gotcha": {
            "type": "string",
            "title": "Critical Gotcha"
          },
          "rotating_item": {
            "type": "string",
            "title": "Rotating Item"
          },
          "quiz_poll": {
            "$ref": "#/components/schemas/QuizPoll"
          }
        },
        "type": "object",
        "required": [
          "one_liner",
          "action_item",
          "critical_gotcha",
          "rotating_item",
          "quiz_poll"
        ],
        "title": "PromptGuideItems"
      },
      "QuizPoll": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "question",
          "options",
          "answer",
          "explanation"
        ],
        "title": "QuizPoll"
      },
      "RelatedNews": {
        "properties": {
          "big_tech": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedNewsItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "industry_biz": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedNewsItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "new_tools": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedNewsItem"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RelatedNews"
      },
      "RelatedNewsItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url",
          "summary"
        ],
        "title": "RelatedNewsItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}